Backend: Implemented User Objects
User Activity now logs if a hint or resource buttonhas been presses once.

User Challenge now calculate points and checks if reflection is completed.
It also checks if an answer is correct according to the solution.

User Reflection checks if an answer is correct. If a solution is empty it award points.
Implemented voting functionality.
Implemented an isCompleted function assissting User Challenge in checking if a reflection is completed.