Backend: Implemented New User Challenge Queries
Implemented add user challenge query.
It first checks if a user is logged in.
It then checks if the challenge exist, before it populates a new user challenge object and sub objects with data.
Then user challenges are filtered on challenge and user id to check if this is the first try fr this challenge or not.
A user challenge awarded max points is also checked for to make sure we do not award the user more points after they have collected all of them.
The points are calculated, the answer is checked, and the reflection is checked for completion before it is stored to the database.

Implemented a delete challenge query.
It checks if user is logged in and is an administrator.
If the input is not null, it checks if a user challenge with that id exists.
The user challenge is then deleted from database.