Backend: Deleted Firestore Objects
Deleted progressess as firestore objects as the progress for each user is calculated and added directly to objects supporting the progress feature.

Deleted userChallenges as firestore object as the user challenge object is added to a user's challenge list and does not need to be a collection on its own.