Backend: Implemented Progress
Removed ID field as a progress is a part of the category or achievement.

Removed updateProgress function as progress only have a single field percentage, and is therefore not necessary.

Implemented the set progress function, taking as an input a percentage, check if it is an integer and assigns it to the percentage field.

Implemented the calculate progress function, taking as an input a user list of user challenges and the list of total challenges or conditions it should calculate progress against.
It then filters out all unique challenges completed and returns a percentage.