Collect and send user challenge data to backend
Implements functionality to capture user data in the various screens
making up a challenge, and then send this data to the backend.

On the backend the answer field is added to the GraphQL type of the user
activity, and the input type for the user challenge is renamed to
UserChallengeInput.