Backend: Implemented Achievement
Implemented remove condition function checking if condition is an array.
It then loops through the array and checks if the condition exists, and removes it from the position.

Implemented check condition function checking if the conditions for an achievement is met.
If all conditions are met the achievement is awarded.

Implemented the convert to response object function returning the format of the response object for an achievement.

Implemented the convert to stored object function converting an achievement object to the format to be stored in database.