Backend: New GraphQL Queries for Achievements
Implemented query to retrieve all achievements.
It gets all achievements, and calculates the individual users progress towards the achievements before all achievements are returned.

Implemented the create achievement query which requires a user to be logged in and that the user has permission of admin.
The achievement is then created, initialized with data, stored to the database and sent as a response object back to the client.

Implemented add user achievement query.
First it checks if user is logged in. It then loops through all ids for achievements given as input which it should add to the user if they have achieved it.
The achievement is initialized and calculated if progess has reached 100% to achieve it.
If the user has completed the achievement it is awarded by storing it in the user's achievement field.
The user achievement is then returned to the client.