Change answer type to string in UserReflection
For challenges where the reflection answer is not a pure string, we need
to serialize the solution. The only other reflection type as of now is
the construction of an argument. In this case the solution will be a
serialized JSON array.

On the frontend we JSON.stringify the answer from the
ArgumentConstructor component.