Remove cachedToken in authLink
This strategy caused a lot of headache when testing rendering of game
elements based on permission levels of users. The problem was that the
cachedToken was not reset on user logout. If a new user logged into the
same device the cached access token from the previous user was used for
subsequent requests.

The idea with keeping a cached copy of the token was to limit the number
of async calls to get the stored access token. We will keep doing async
calls until a new strategy is implemented.