Backend: Readying Express Server for Subscription
Server file have been modified to add the pubsub functionality to the graphql context.
It makes the pubsub available to all endpoints.

A new web socket server is created to make the connection persistent when a user subscribes.
The web socket is then integrated with the express server by the useServer function.

A new express middelware is created to handle calls to /graphiql, returning a new website for interacting with GraphiQL supporting websockets and allowing GraphQL to have subscriptions.

The new website for GraphiQL is found in backend root folder and is called graphiql-over-ws.html.
Thanks to  Github user enisdenjo for the html.
It could be found at:
https://gist.github.com/enisdenjo/a68312878fdc4df299cb0433c60c1dea

Utility file for GraphQL now import GraphQL Redis Subscription dependency and exports a new redis pub sub object.

Made changes to route.js making the GraphQL Schema supporting subscriptions.