Backend: Refactor Progress
Progress is removed from category and achievement objects as they no longer keep track of individual users progress.

Instead the progress is relocated to remain in the user*s data.
The user now have field for progress in its data, containing progress for categories and achievements.

The category query no longer calculates individual users progress as the progress is calculated when a user challenge is completed instead.
The same reasoning holds when getting achievements, it no longer calculates individual users progress.

Progress for achievement and category is now calculated when  user completes a challenge and a user challenge is added to the user's profile.
Adding a user challenge also publish a update user profile message to users subscribed to updateUserProfile endpoint.
The endpoint must be restricted to only return data to the correct user.

As the user now stores the progress, the model have been updated to support that functionality.
The user model now returns a progress with progress for category and achievement.

Removed password field from user model for security reasons.