Backend: Added User Specific Subscription
When a user subscribe to userChallengeAdded, the user with the correct access token as input to subscription is the only user to recieve the request.

This is only a proof of concept as access token expires and it forces front end to update subscription on expiration. It is therefore not a valid solution.
Consider creating a subscribe token to be stored in database and deleted on log out.