Backend: Refactor Token, User Model Update
Refactor the creation of refresh tokens in accessToken.js
The store refresh token in database function, along with create refresh token function have been deleted and replaced with create token function and store in database function.
The new functions handle creation and storage of refresh tokens and subscription tokens.

Sign in now use the new create token function to create refresh and subscription tokens.

The sign in model is updated to provide a field for subscribe token.