Backend: Add User Challenge Gives Achievement
Updated add user challenge query to calculate progress for achievement and award it if the achievement is completed.
Achievements already awarded is not added twice.

Deleted add user achievement query as it is the add user challenge query who should award the achievement.
Route for add user achievement has been removed as an endpoint for GraphQL.