Implement initial achievement screen
All available achievements are currently rendered in a FlatList with
four columns. An avatar icon is used to represent an achievement if no
image is set for it. The background color for an achievement is
generated based on the achievement ID. Detailed information about an
achievement are rendered in a Modal when an item in the grid is touched.
The achievement list is also wrapped in a native stack to get a
consistent screen header view for the application.