Backend: Implemented Leaderboard
Implemented the leaderboard functionalities as setting a name for the leaderboard, adding users to the leaderboard, removing users from the leaderboard, calculating a highscore leaderboard, calculate a leaderboard per category, and calculate leaderboards for different difficulties.

Implemented get points functionality for a profile, calculating the points the profile has aquired based on a filter and a criteria.
If filter and criteria is not set, a highcore for all challenges combined are calculated.

Implemented a query retrieveing the leaderboard if the user is authenticated and has permission of experiment.
A highscore leaderboard and a leaderboard for category is available.

Implemented a GraphQL model for leaderboard response data.

Added a new endpoint to GraphQL for retrieving the leaderboard.

Leaderboard query is now exported by internal.js and made available to other files.