Fetch complete userProfile in ContentProvider
The signIn and activateAccount mutations in the account package no only
request bare user data. The complete userProfile data is delegated to
the ContentProvider. This way we won't hit the size limit for values
provided to SecureStore, as described in issue #148.

On the backend a new userProfile query and resolver is added. This is
essentially a rename of getUserByID with the addition of a check that
the request is authenticated.

Closes #148