Backend: New Leaderboards and Subscription
Leaderboard now calculates points for category and difficulty.

Removed delete leaderboard function and calculating leaderboard for difficulty within a category.

Updated the stored user challenge object with difficulty and category id as the leaderboard needs it as criteria to filter challenges on.

When retrieving the leaderboard, the leaderboard per category and difficulty is also provided in the response.

When adding a new user challenge, the leaderboard calculates the new score for the user that completed the challenge, and broadcast their new score to all users who has subscribed to the leaderboard.

New leaderboard data models for category data and difficulty data.

Added new subscription endpoint for leaderboard to GraphQL.