Setup GraphQL subscription on frontend
The ContentProvider now subscribes to userChallengeAdded, which triggers
when a user challenge is added. This way the rest of the UI is updated
with fresh userProfile data upon completing a challenge.

A new environment variable is introduced with SUBSCRIPTION_ENDPOINT
which should point towards the WebSocket URL on the backend. For
consistency the API_ENDPOINT variable is renamed to HTTP_ENDPOINT. The
README.md is updated to reflect this changes.

The frontend no also receives a subscribeToken which is used to
authenticate subscription connections. This variable is handled by the
AuthProvider.