Update logic for creating leaderboards
On the backend GraphQL type definitions for the leaderboard resolver are
updated, and now returns several high score lists:

- one all time high score list,
- one high score list for each category, and
- one high score list for each difficulty level.

Further more, the leaderboard module is completely rewritten. The rank
for each user in a high score list is now calculated.

On the frontend both queries for and rendering of leaderboards are
updated accordingly.

Closes #47