Refactor leaderboards resolver and subscription
The actual high score rank annotation is now delegated to the frontend.
This should be as simple as sorting the scores and using index + 1 for
items in the list. The reason for this is to simplify the subscription
link where we only publish partial high scores. Delegating ranking to
the frontend means that we don't have to recreate the complete
leaderboard for every subscription.

Instead of the rank the userID is now returned as part of the high score
list entry.

The high scores for categories and difficulty are removed for the time
being. They should be fairly easy to implement again but we won't
prioritize this for this project.