Backend: SmileOMeter Functionality
Added new firestore provider for smile-o-meter.

Added new query for logging smile-o-meter data in Likert scale format per challenge per user.
It takes as input an evaluation for a challenge from a user, consisting of a challenge id and a 5-point  Likert Scale value.
It checks if the user is authenticated and if the challenge to be evaluated exists.
The smile-o-meter data is retrieved for a challenge and it logs the evaluation from the user to the smile-o-meter data for that challenge. It also accounts for the user's permission when performing the evaluation to differentiate feedback on permission.

Implemented a amile-o-meter input model for GraphQL using a Likert Scale enum model when recieving values from the smile-o-meters.

Added a new endpoint for GraphQL named logSmileOMeter, and should be used by frontend to log smile-o-meter evaluations from users.

Added query file for smile-o-meter to export in internal.js.