Remove calculation of points on frontend
In order to provide consitent feedback on rewards the calculation of
points is now handled only by the backend. This way we can be sure that
we provide consistent feedback on user rewards both when completing a
challenge and when updating the leaderboards.

Also added some fade to the HeroBackgroundImage and moved the ScrollView
inside it.