Fix error trying to trim null value
The error happens when frontend reloads on the Completed screen tries to
add a userChallenge with dateCompleted set to null. It is rather an edge
case that most likely will never happen in prod, but never the less an
extra check is added and a proper API error is thrown when the
dateCompleted value is null.