Add app icons
This commit add custom app icons for iOS and Android and a matching
splash screen. A Python script to export the various icons from the
included SVG source file is also implemented. The app.json file is
updated so the backgroundColor for both android.adaptiveIcon, splash,
and androidStatusBar matches the primaryColor.

Closes #217