Refactor meta chip rendering
Because Chip components are rendered the same in both the challenge list
and the challenge intro screen, the rendering of these meta chips is
moved to its own Chip module inside the challenge feature.

Also the difficulty "enum" declaration as well as the code to select
colors for difficulty levels are moved to the shared folder.